Commit changes for separating the modules related to todo items into appropriate components
